// types.ts
export enum GuaranteeType {
  VEHICLE = 'VEHICLE',
  PROPERTY = 'PROPERTY',
  OTHER = 'OTHER'
}
// ...
export interface Guarantee {
  id: string;
  user_id: string;
  tipo_garantia: GuaranteeType;
  descricao: string;
  valor_estimado: number;
  deseja_vender: boolean;
  valor_venda?: number;
  status: 'PENDING' | 'VERIFIED' | 'REJECTED';
  fotos?: string[];
  // Campos específicos de Veículo
  marca?: string;
  modelo?: string;
  ano?: number;
  placa?: string;
  // Campos específicos de Imóvel
  tipo_imovel?: string;
  area_total?: number;
  matricula?: string;
}